+++ src/system/kernel/debugger/SerialIO.cc
@@ -174,11 +174,11 @@ void SerialIO::drawHorizontalLine(char c, size_t row, size_t colStart, size_t co
 
   if (colEnd >= m_nWidth)
     colEnd = m_nWidth-1;
-  if (colStart < 0)
+  if (static_cast<int32_t>(colStart) < 0)
     colStart = 0;
   if (row >= m_nHeight)
     row = m_nHeight-1;
-  if (row < 0)
+  if (static_cast<int32_t>(row) < 0)
     row = 0;
 
   startColour(foreColour, backColour);
@@ -246,11 +246,11 @@ void SerialIO::drawVerticalLine(char c, size_t col, size_t rowStart, size_t rowE
 
   if (rowEnd >= m_nHeight)
     rowEnd = m_nHeight-1;
-  if (rowStart < 0)
+  if (static_cast<int32_t>(rowStart) < 0)
     rowStart = 0;
   if (col >= m_nWidth)
     col = m_nWidth-1;
-  if (col < 0)
+  if (static_cast<int32_t>(col) < 0)
     col = 0;
   
   // TODO position cursor, draw line.
